Add tests for delete_samples() edge cases and consolidate delete tests
The previously untested delete_samples() paths include empty list is a noop, None raises TypeError, and a non-existent sample name raises RuntimeError.
