+++ apis/python/tests/test_delete.py
@@ -1,0 +1,73 @@
+import os
+
+import pytest
+import tiledbvcf
+
+from .conftest import skip_if_no_bcftools, TESTS_INPUT_DIR
+
+
+def test_delete_dataset(tmp_path):
+    uri = os.path.join(tmp_path, "delete_dataset")
+
+    with tiledbvcf.Dataset(uri, mode="w") as ds:
+        ds.create_dataset()
+
+    # Check that the dataset exists
+    assert os.path.exists(uri)
+
+    # Delete the dataset
+    tiledbvcf.Dataset.delete(uri)
+
+    # Check that the dataset does not exist
+    assert not os.path.exists(uri)
+
+
+@skip_if_no_bcftools
+def test_delete_samples(tmp_path, stats_v3_dataset, stats_sample_names):
+    assert "second" in stats_sample_names
+    assert "fifth" in stats_sample_names
+    assert "third" in stats_sample_names
+    ds = tiledbvcf.Dataset(uri=os.path.join(tmp_path, "stats_test"), mode="w")
+    # tiledbvcf.config_logging("trace")
+    ds.delete_samples(["second", "fifth"])
+    ds = tiledbvcf.Dataset(uri=os.path.join(tmp_path, "stats_test"), mode="r")
+    sample_names = ds.samples()
+    assert "second" not in sample_names
+    assert "fifth" not in sample_names
+    assert "third" in sample_names
+
+
+def test_delete_samples_empty_list_is_noop(tmp_path):
+    """delete_samples([]) leaves all samples in the dataset untouched."""
+    uri = os.path.join(tmp_path, "dataset")
+    ds = tiledbvcf.Dataset(uri, mode="w")
+    ds.create_dataset()
+    ds.ingest_samples([os.path.join(TESTS_INPUT_DIR, s) for s in ["small.bcf", "small2.bcf"]])
+
+    ds = tiledbvcf.Dataset(uri, mode="w")
+    ds.delete_samples([])
+
+    ds = tiledbvcf.Dataset(uri, mode="r")
+    assert set(ds.samples()) == {"HG00280", "HG01762"}
+
+
+def test_delete_samples_none_raises(tmp_path):
+    """delete_samples(None) raises because the C++ writer requires a sequence."""
+    uri = os.path.join(tmp_path, "dataset")
+    ds = tiledbvcf.Dataset(uri, mode="w")
+    ds.create_dataset()
+
+    with pytest.raises(TypeError):
+        ds.delete_samples(None)
+
+
+def test_delete_samples_nonexistent_raises(tmp_path):
+    """Attempting to delete a sample that is not in the dataset raises RuntimeError."""
+    uri = os.path.join(tmp_path, "dataset")
+    ds = tiledbvcf.Dataset(uri, mode="w")
+    ds.create_dataset()
+    ds.ingest_samples([os.path.join(TESTS_INPUT_DIR, "small.bcf")])
+
+    ds = tiledbvcf.Dataset(uri, mode="w")
+    with pytest.raises(RuntimeError, match="Sample not found in dataset"):
+        ds.delete_samples(["NONEXISTENT"])
